+++ src/components/+slidePanel.svelte
@@ -1,0 +1,16 @@
+<script lang="ts">
+    import { countryContentStore } from '../store/mapStore';
+
+    let countryContent: string = '';
+
+    countryContentStore.subscribe(value => {
+        countryContent = value;
+    });
+</script>
+
+
+<div class ="flex">
+    <aside id="slide-panel" class="flex max-h-full fixed h-full items-center max w-1/5 right-0 top-0 justify-between px-6 bg-white text-gray-700 border-b border-gray-200 z-10">
+      {countryContent}
+    </aside>
+</div>
