+++ src/routes/+page.svelte
@@ -1,19 +1,17 @@
 <script lang="ts">
     import { onMount } from 'svelte';
-    import { tooltipContent, handleFormSubmit, setupMapInteractions, initializeCountryMap, countries} from '../scripts/mapInteractions';
-    import { countryStore, countryContentStore, tooltipToggler, sidepanelToggler, xStore, yStore } from '../store/mapStore';
+    import { tooltipContent, setupMapInteractions, initializeCountryMap} from '../scripts/mapInteractions';
+    import { tooltipToggler, sidepanelToggler, xStore, yStore } from '../store/mapStore';
+    import Nav from '../components/+nav.svelte';
+    import SlidePanel from '../components/+slidePanel.svelte';
     import { zoom, type ViewBox } from '../scripts/zoom';
-
     export let data;
-
-    let svgElement: SVGSVGElement;
+    
+    let svgElement : SVGSVGElement;
     let viewBox: ViewBox = { x: 0, y: 0, width: 2000, height: 857 };
-
-    let selectedCountry : string = $countryStore;
     let tooltipVisible: boolean = false;
     let tooltipX: number = 0;
     let tooltipY: number = 0;
-    let countryContent: string = '';
     let showSidePanel: boolean = false;
 
     onMount(() => {
@@ -29,21 +27,12 @@
 
     //Store variables
 
-    function updateCountry(event: Event) {
-        const input = event.target as HTMLInputElement;
-        countryStore.set(input.value);
-    }
-
     tooltipToggler.subscribe(value => {
         tooltipVisible = value;
     });
 
     sidepanelToggler.subscribe(value => {
         showSidePanel = value;
-    });
-
-    countryContentStore.subscribe(value => {
-        countryContent = value;
     });
 
     xStore.subscribe(value => {
@@ -58,29 +47,11 @@
 </script>
 
 <div class="flex flex-col w-100 h-auto max-h-full overflow-hidden overscroll-contain items-center justify-center bg-[#5cb5e1] backdrop-blur-lg">
-    <nav class="w-full h-20 lg:h-24 grid grid-cols-3 bg-[#323638]">
-        <div class="flex justify-start pl-16 items-center m-0 font-dosis text-white text-2xl max-h-full gap-2">
-            <img src="/logo.png" alt="logo" class="max-w-full max-h-full object-contain h-[75px]">
-            reKnewable
-        </div>
-        <div class="flex justify-center items-center m-0 text-white text-4xl font-dosis">
-            <form class="relative w-full flex max-h-full" on:submit="{handleFormSubmit}">
-                <input type="search" bind:value={selectedCountry} on:input="{updateCountry}" class="block w-full p-4 pl-10 text-sm text-gray-900 border border-gray-300 rounded-lg bg-gray-50 outline-none font-dosis" placeholder="Search countries, continents..." required />
-                <button type="submit" class="absolute border-2 right-2 top-2 text-white bg-[#323638] hover:opacity-[75%] focus:ring-4 focus:outline-none font-dosis rounded-lg text-sm px-4 py-2 ">Search</button>
-            </form>
-        </div>
-        <div class="flex justify-end pr-16 items-center m-0 font-dosis text-white text-2xl">
-            Menu
-        </div>
-    </nav>
+    <Nav></Nav>
     <div class="flex items-center justify-center bg-[#5cb5e1] w-full h-screen">
         {#if showSidePanel}
-        <div class ="flex">
-          <aside id="slide-panel" class="flex max-h-full fixed h-full items-center max w-1/5 right-0 top-0 justify-between px-6 bg-white text-gray-700 border-b border-gray-200 z-10">
-            {countryContent}
-          </aside>
-        </div>
-      {/if}
+            <SlidePanel></SlidePanel>
+        {/if}
       <figure on:wheel={handleWheel}>
         <svg bind:this={svgElement} baseProfile="tiny" fill="#ececec" height="857" stroke="black" stroke-linecap="round" stroke-linejoin="round" stroke-width=".2" version="1.2" viewBox="0 0 2000 857" width="2000" xmlns="http://www.w3.org/2000/svg">
             <circle cx="997.9" cy="189.1" id="0">
